+++ app/Http/Traits/ModelTrait.php
@@ -1,0 +1,163 @@
+<?php
+
+namespace App\Http\Traits;
+
+use App\Exceptions\RedirectException;
+use Illuminate\Database\Eloquent\Model;
+use Illuminate\Database\Eloquent\ModelNotFoundException;
+use Illuminate\Http\Request;
+
+/**
+ * Class ModelTrait
+ *
+ * @author EB
+ * @package App\Http\Traits
+ */
+trait ModelTrait
+{
+    /**
+     * @author WN
+     * @param Model $model
+     * @param int $id
+     * @param string $modelName
+     * @param string $redirect
+     * @return Model
+     * @throws RedirectException
+     */
+    protected function fetchModelById(Model $model, $id, $modelName, $redirect)
+    {
+        try {
+            return $model->findOrFail($id);
+
+        } catch (ModelNotFoundException $e) {
+
+            $this->logError(
+                'Could not get ' . $modelName . ' with ID [' . $id . ']; ' .
+                ucwords($modelName) . ' does not exist: ' . $e->getMessage()
+            );
+            throw (new RedirectException())
+                ->setTarget($redirect)
+                ->setError('Could not found ' . ucwords($modelName) . ' with ID:' . $id);
+        }
+    }
+
+    /**
+     * @author WN
+     * @param Model $model
+     * @param int $id
+     * @param string $modelName
+     * @param string $redirect
+     * @return \Illuminate\Http\RedirectResponse
+     * @throws RedirectException
+     */
+    protected function destroyModel(Model $model, $id, $modelName, $redirect)
+    {
+        try {
+            $model->destroy($id);
+
+        } catch (ModelNotFoundException $e) {
+
+            $this->logError('Deletion of this record did not complete successfully' . $e->getMessage());
+            throw (new RedirectException())
+                ->setTarget($redirect)
+                ->setError('Deletion of this record did not complete successfully');
+        }
+
+        return redirect($redirect)->with('messages', ['success', ucwords($modelName) . ' was successfully deleted']);
+    }
+
+    /**
+     * @author WN
+     * @param Model $model
+     * @param int $id
+     * @param string $modelName
+     * @param string $redirect
+     * @param Request $request
+     * @return \Illuminate\Http\RedirectResponse
+     * @throws RedirectException
+     */
+    protected function updateModel(Model $model, $id, $modelName, $redirect, Request $request)
+    {
+        $model = $this->fetchModelById($model, $id, $modelName, $redirect);
+        try{
+            $this->updateActiveField($model, $request->has('active'));
+            $model->update($request->all());
+        } catch (\Exception $e) {
+
+            throw (new RedirectException())->setTarget($redirect . '/' . $id . '/edit')->setError($e->getMessage());
+        }
+        return redirect()->back()->with('messages', ['success' => ucwords($modelName) .' details were successfully updated']);
+    }
+
+    /**
+     * @author EB, WN
+     * @param Model $model
+     * @param bool $active
+     * @return Model
+     */
+    protected function updateActiveField($model, $active)
+    {
+        if ($model->active xor $active) {
+            if ($active) {
+                if (method_exists($model, 'activate')) {
+                    $model->activate();
+                }
+            } else {
+                if (method_exists($model, 'deactivate')) {
+                    $model->deactivate();
+                }
+            }
+        }
+
+        return $model;
+    }
+
+    /**
+     * @author WN
+     * @param Model $entity
+     * @param int $merchantId
+     * @param string $redirect
+     * @param string $modelName
+     * @return Model
+     * @throws RedirectException
+     */
+    protected function checkModelForMerchantLimit(Model $entity, $merchantId, $modelName, $redirect)
+    {
+        if (!$this->isMerchantAllowedForUser($merchantId)) {
+            throw RedirectException::make($redirect)
+                ->setError('You are not allowed to take any action on this' . ucwords($modelName));
+        }
+
+        return $entity;
+    }
+
+    /**
+     * @author WN
+     * @param int $merchantId
+     * @return bool
+     */
+    protected function isMerchantAllowedForUser($merchantId)
+    {
+        if (empty($this->getAuthenticatedUser()->merchant_id) ||
+            $this->getAuthenticatedUser()->merchant_id == $merchantId
+        ) {
+            return true;
+        }
+
+        return false;
+    }
+
+    /**
+     * @author WN
+     * @param string $message
+     * @param array $context
+     * @return null
+     */
+    abstract function logError($message, array $context = array());
+
+    /**
+     * @author WN
+     * @return \App\User|null
+     */
+    abstract protected function getAuthenticatedUser();
+}
